Changed Index.js and changed layout
I made the index detect keywords like lag or echo and then make a list of responses. The responses might not show up, but in the js console it does. In order to access the console right click anywhere on the page and then click inspect element. There should be a tab that says console.